+++ src/Functions/Math/Divide.php
@@ -8,25 +8,17 @@ use FQL\Functions\Core\MultipleFieldsFunction;
 
 final class Divide extends MultipleFieldsFunction
 {
-    public function __construct(string|float|int ...$fields)
-    {
-        parent::__construct(
-            ...array_map(fn($field) => (is_float($field) || is_int($field)) ? sprintf('"%s"', $field) : $field, $fields)
-        );
-    }
-
     public function __invoke(array $item, array $resultItem): mixed
     {
         $acc = null;
         foreach ($this->fields as $field) {
-            $value = $this->getFieldValue($field, $item, $resultItem);
+            $value = $this->getFieldValue($field, $item, $resultItem) ?? $field;
+            if (is_string($value)) {
+                $value = Type::matchByString($value);
+            }
 
             if ($value === null || $value === '') {
                 $value = 0;
-            }
-
-            if (is_string($value)) {
-                $value = Type::matchByString($value);
             }
 
             if (!is_numeric($value) && is_string($value)) {