+++ src/Interface/Query.php
@@ -351,10 +351,10 @@ interface Query extends \Stringable
      */
     public function modulo(string $field, int $divisor): Query;
 
-    public function add(string|float|int ...$fields): Query;
-    public function subtract(string|float|int ...$fields): Query;
-    public function multiply(string|float|int ...$fields): Query;
-    public function divide(string|float|int ...$fields): Query;
+    public function add(string ...$fields): Query;
+    public function subtract(string ...$fields): Query;
+    public function multiply(string ...$fields): Query;
+    public function divide(string ...$fields): Query;
 
     /**
      * Apply the LENGTH function to a field in the query.