+++ tests/Functions/Math/SubTest.php
@@ -10,7 +10,7 @@ class SubTest extends TestCase
 {
     public function testSub(): void
     {
-        $sub = new Sub('price', 5);
+        $sub = new Sub('price', "5");
         $this->assertEquals(5, $sub(['price' => 10], []));
     }
 
@@ -18,16 +18,22 @@ class SubTest extends TestCase
     {
         $sub = new Sub('price', '"2"');
         $this->assertEquals(3, $sub(['price' => 5], []));
+        $sub = new Sub('price', '2');
+        $this->assertEquals(3, $sub(['price' => 5], []));
     }
 
     public function testSubWithEmptyValue(): void
     {
         $sub = new Sub('price', '"2"');
         $this->assertEquals(-2, $sub(['price' => ''], []));
+
+        $sub = new Sub('price', '2');
+        $this->assertEquals(-2, $sub(['price' => ''], []));
     }
 
     public function testSubWithNullValue(): void
     {
+        $this->expectException(UnexpectedValueException::class);
         $sub = new Sub('price', '"2"');
         $this->assertEquals(-2, $sub(['price' => null], []));
     }
@@ -35,6 +41,9 @@ class SubTest extends TestCase
     public function testSubWithFloat(): void
     {
         $sub = new Sub('price', '"2.5"');
+        $this->assertEquals(97.5, $sub(['price' => 100], []));
+
+        $sub = new Sub('price', '2.5');
         $this->assertEquals(97.5, $sub(['price' => 100], []));
     }
 }